根据代码生成生成多个数据库操作，修改码表的list分页到查全部，

+++ src/main/java/org/ml/blog/service/SysCodeService.java
@@ -1,18 +1,16 @@
 package org.ml.blog.service;
 
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
 import org.ml.blog.common.ResultCode;
 import org.ml.blog.domain.entity.SysCodeType;
 import org.ml.blog.domain.entity.SysCodeValue;
-import org.ml.blog.domain.form.SysCodeTypeForm;
-import org.ml.blog.domain.form.SysCodeTypeUpdateForm;
-import org.ml.blog.domain.form.SysCodeValueForm;
-import org.ml.blog.domain.form.SysCodeValueUpdateForm;
+import org.ml.blog.domain.dto.SysCodeTypeDTO;
+import org.ml.blog.domain.dto.SysCodeValueDTO;
+import org.ml.blog.domain.dto.SysCodeValueUpdateDTO;
 import org.ml.blog.domain.vo.PageVO;
 import org.ml.blog.domain.vo.SysCodeTypeVO;
 import org.ml.blog.domain.vo.SysCodeValueVO;
 import org.ml.blog.exception.BizException;
+import org.ml.blog.repository.RedisRepository;
 import org.ml.blog.repository.SysCodeTypeRepository;
 import org.ml.blog.repository.SysCodeValueRepository;
 import org.ml.blog.utils.MyBeanUtils;
@@ -24,6 +22,7 @@ import org.springframework.cache.annotation.Caching;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
@@ -37,7 +36,10 @@ public class SysCodeService {
     SysCodeValueRepository sysCodeValueRepository;
     @Autowired
     SysCodeTypeRepository sysCodeTypeRepository;
-
+    @Autowired
+    RedisTemplate<String,Object>  redisTemplate;
+    @Autowired
+    RedisRepository redisRepository;
     @Cacheable(value = "sys_type", key = "#codeType")
     public List<SysCodeValueVO> queryCode(String codeType) {
         SysCodeType type = sysCodeTypeRepository.findByCodeType(codeType);
@@ -51,22 +53,14 @@ public class SysCodeService {
 
 
     @Cacheable(value = "sys_code_type", key = "'list'")
-    public PageVO<SysCodeTypeVO> queryList(int page, int size) {
-        Pageable pageable = PageRequest.of(page, size);
-        Page<SysCodeType> all = sysCodeTypeRepository.findAll(pageable);
-
-        List<SysCodeTypeVO> list = all.stream().map(e -> {
-            SysCodeTypeVO vo = new SysCodeTypeVO();
-            BeanUtils.copyProperties(e, vo);
-            return vo;
-        }).toList();
-        return PageVO.getPageVO(all, list);
-
+    public List<SysCodeTypeVO> list() {
+        List<SysCodeType> list= sysCodeTypeRepository.findAll();
+        return MyBeanUtils.copyList(list, SysCodeTypeVO::new);
     }
 
     // 删除码表list的缓存
     @CacheEvict(value = "sys_code_type", key = "'list'")
-    public SysCodeTypeVO insertSysCodeType(SysCodeTypeForm form) {
+    public SysCodeTypeVO insertSysCodeType(SysCodeTypeDTO form) {
         SysCodeType sysCodeType = new SysCodeType();
         BeanUtils.copyProperties(form, sysCodeType);
         SysCodeType byCodeType = sysCodeTypeRepository.findByCodeType(form.getCodeType());
@@ -84,7 +78,7 @@ public class SysCodeService {
             @CacheEvict(value = "sys_code_type", key = "'list'"),
             @CacheEvict(value = "sys_type", key = "#form.codeType")
     })
-    public SysCodeTypeVO updateSysCodeType(SysCodeTypeForm form) {
+    public SysCodeTypeVO updateSysCodeType(SysCodeTypeDTO form) {
         SysCodeType byCodeType = sysCodeTypeRepository.findByCodeType(form.getCodeType());
         if (byCodeType == null) {
             throw new BizException(ResultCode.DATA_NO_EXISTS);
@@ -113,7 +107,7 @@ public class SysCodeService {
     }
 
     @CacheEvict(value = "sys_type", key = "#form.codeType")
-    public SysCodeValueVO insertSysCodeValue(SysCodeValueForm form) {
+    public SysCodeValueVO insertSysCodeValue(SysCodeValueDTO form) {
         SysCodeType byCodeType = sysCodeTypeRepository.findByCodeType(form.getCodeType());
         if (byCodeType == null) {
             throw new BizException(ResultCode.DATA_NO_EXISTS);
@@ -132,7 +126,7 @@ public class SysCodeService {
 
 
     @CacheEvict(value = "sys_type", key = "#form.codeType")
-    public SysCodeValueVO updateSysCodeValue(SysCodeValueUpdateForm form) {
+    public SysCodeValueVO updateSysCodeValue(SysCodeValueUpdateDTO form) {
         SysCodeValue sysCodeValueById = sysCodeValueRepository.findSysCodeValueById(form.getId());
         if (sysCodeValueById == null) {
             throw new BizException(ResultCode.DATA_NO_EXISTS);
@@ -145,9 +139,17 @@ public class SysCodeService {
 
     }
 
-    @CacheEvict(value = "sys_type", key = "#codeType")
     @Transactional
-    public void delSysCodeValue(String codeType, String codeValueId) {
+    public void delSysCodeValue(String codeValueId) {
+        // 删除缓存
+        SysCodeValue sysCodeValue = sysCodeValueRepository.findSysCodeValueById(codeValueId);
+        if (sysCodeValue==null) {
+            throw new BizException(ResultCode.DATA_NO_EXISTS);
+        }
+        redisTemplate.delete("sys_type::" + sysCodeValue.getCodeValue());
+        // 写数据库
         sysCodeValueRepository.deleteSysCodeValueById(codeValueId);
+        // 延迟删除缓存
+        redisRepository.asyncDelete("sys_type::" + sysCodeValue.getCodeValue());
     }
 }
